stop_machine: pass curstate to ack_state()
The multi_cpu_stop() state machine uses multi_stop_data::state to hold
the current state, and this is read and written atomically except in
ack_state(), which performs a non-atomic read.

As ack_state() only performs this non-atomic read when there is a single
writer, this is benign, but it makes reasoning about the state machine a
little harder.

Remove the non-atomic read and pass the (atomically read) curstate in
instead. This makes it clear that we do not expect any racy writes, and
avoids a redundant load.

Acked-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Rong Tao <rongtao@cestc.cn>